Repair supported locomotion, retained facing and guarded side kicks
Unify gait/support speed and conserve horizontal task-motor momentum. Fit stance travel to leg reach. Shorten punch recovery, add bounded fist/forearm contact weight, and make the heel path and turned shoulder guard authoritative. Anchor kick height to its support surface and remove the unused ambience call.

Integrates the current anatomy extraction and support yaw wrench. Adds causal severance/replay probes and records measured results and remaining support-foot compliance in EDGES.md.

Verified: test:substrate, typecheck, canonical standalone build, desktop kick/recovery inspection and mobile pointer-stick release checks.

+++ src/game/human-root-authority.ts
@@ -44,7 +44,7 @@ export class HumanRootAuthority {
   private readonly x = new Float32Array(ENTITY_ID_CAP);
   private readonly y = new Float32Array(ENTITY_ID_CAP);
   private readonly z = new Float32Array(ENTITY_ID_CAP);
-  private readonly yaw = new Float32Array(ENTITY_ID_CAP);
+  private readonly yaw = new Float64Array(ENTITY_ID_CAP);
   private readonly valid = new Uint8Array(ENTITY_ID_CAP);
 
   capture(w: World) {
@@ -96,3 +96,4 @@ export class HumanRootAuthority {
     this.valid.fill(0);
   }
 }
+
